Decide death reason when logging
